Removed initialisation of FRONTIER_LOG_FILE environment variable with empty string. Initialising the FRONTIER_LOG_FILE environment variable with empty string is a bug in the most recent versions of the frontier client: you get an error message like:
Cannot open log file . Log is disabled.

Hence, you have no log messages at all.
I checked this both online on the DQM playback and on Hilton, and offline with a recent CMSSW.

By removing the environment variable, the log messages go to stdout, and they are captured in the log file.