Modify startDqmRun.sh to initialize the cmssw environment from the appliance link.
Previously it used the 'current' link in local home directory.

The importance of this change is that it allows us to run multiple cmssw environments at the same time.
Our policy was/is to always run one release area for all the applications, but in very special and rare cases (now), we don't want one application to block upgrades for the rest.